Corregir pequeños errores en colocacion de paquetes
