--- src/Dominio/Camarero.java
@@ -1,22 +1,0 @@
-package Dominio;
-
-public class Camarero {
-
-	private int id_camarero;
-	private String nombre;
-	
-	public int getId_camarero() {
-		return id_camarero;
-	}
-	public void setId_camarero(int id_camarero) {
-		this.id_camarero = id_camarero;
-	}
-	public String getNombre() {
-		return nombre;
-	}
-	public void setNombre(String nombre) {
-		this.nombre = nombre;
-	}
-	
-
-}
