--- src/Dominio/Cliente.java
@@ -1,27 +1,0 @@
-package Dominio;
-
-public class Cliente {
-
-	private String nombre;
-	private int telefono;
-	private int id_cliente;
-	public String getNombre() {
-		return nombre;
-	}
-	public void setNombre(String nombre) {
-		this.nombre = nombre;
-	}
-	public int getTelefono() {
-		return telefono;
-	}
-	public void setTelefono(int telefono) {
-		this.telefono = telefono;
-	}
-	public int getId_cliente() {
-		return id_cliente;
-	}
-	public void setId_cliente(int id_cliente) {
-		this.id_cliente = id_cliente;
-	}
-
-}
