--- src/Persistencia/BBDD.java
@@ -1,4 +1,0 @@
-package Persistencia;
-
-public class BBDD {
-}
