--- src/Presentacion/IU_Jefe_de_Sala.java
@@ -1,16 +1,0 @@
-package Presentacion;
-
-public class IU_Jefe_de_Sala {
-
-	/**
-	 * 
-	 * @param user
-	 * @param pass
-	 */
-	public boolean autentificacion(int user, int pass) {
-		// TODO - implement IU_Jefe_de_Sala.autentificacion
-		
-		return false;
-	}
-
-}
