+++ src/main/java/Dominio/Cliente.java
@@ -2,28 +2,26 @@ package Dominio;
 
 public class Cliente {
 
-    private String nombre;
-    private int telefono;
-
-    public Cliente(String nombre, int telefono) {
-        this.nombre = nombre;
-        this.telefono = telefono;
-    }
-
-    public void setNombre(String nombre) {
-        this.nombre = nombre;
-    }
-
-    public void setTelefono(int telefono) {
-        this.telefono = telefono;
-    }
-
-    public String getNombre() {
-        return nombre;
-    }
-
-    public int getTelefono() {
-        return telefono;
-    }
+	private String nombre;
+	private int telefono;
+	private int id_cliente;
+	public String getNombre() {
+		return nombre;
+	}
+	public void setNombre(String nombre) {
+		this.nombre = nombre;
+	}
+	public int getTelefono() {
+		return telefono;
+	}
+	public void setTelefono(int telefono) {
+		this.telefono = telefono;
+	}
+	public int getId_cliente() {
+		return id_cliente;
+	}
+	public void setId_cliente(int id_cliente) {
+		this.id_cliente = id_cliente;
+	}
 
 }
