+++ src/main/java/Dominio/gestor_Jefe_de_Sala.java
@@ -2,19 +2,19 @@ package Dominio;
 
 public class gestor_Jefe_de_Sala {
 
-    public void asignarCamarero() {
-        // TODO - implement gestor_Jefe_de_Sala.asignarCamarero
+	public void asignarCamarero() {
+		// TODO - implement gestor_Jefe_de_Sala.asignarCamarero
+		
+	}
 
-    }
+	public void asignarMesa() {
+		// TODO - implement gestor_Jefe_de_Sala.asignarMesa
+		
+	}
 
-    public void asignarMesa() {
-        // TODO - implement gestor_Jefe_de_Sala.asignarMesa
-
-    }
-
-    public void gestion_estados_mesa() {
-        // TODO - implement gestor_Jefe_de_Sala.gestion_estados_mesa
-
-    }
+	public void gestion_estados_mesa() {
+		// TODO - implement gestor_Jefe_de_Sala.gestion_estados_mesa
+		
+	}
 
 }
