+++ src/main/java/Presentacion/IU_Jefe_de_Sala.java
@@ -1,22 +1,16 @@
 package Presentacion;
 
-import Dominio.*;
-
 public class IU_Jefe_de_Sala {
 
-    public Camarero asignar_camarero() {
-        return null;
-
-    }
-
-    public Mesa asignar_mesa() {
-        return null;
-
-    }
-
-    public void gestion_estados_mesa() {
-        // TODO - implement IU_Jefe_de_Sala.gestion_estados_mesa
-
-    }
+	/**
+	 * 
+	 * @param user
+	 * @param pass
+	 */
+	public boolean autentificacion(int user, int pass) {
+		// TODO - implement IU_Jefe_de_Sala.autentificacion
+		
+		return false;
+	}
 
 }
